♻ refactor: 디렉토리명 변경, 주석 삭제
